Adding repo line to layer.yaml
If the charm is built as-is, its generated layer.yaml will end up with
the line:

  "repo": "https://git.launchpad.net/layer-snap"

With this tweak, the repo saved into layer.yaml will instead read:

  "repo": "https://code.launchpad.net/charm-juju-local"

This delta was observed between cs:~peter-sabaini/charm-juju-local-1
and cs:~canonical-bootstack/juju-local-0, where it appears the repo
line was hand-tweaked in the latter.  This change will preserve that
change for future builds.