charm: install lxd snap
In the ubuntu charms with series bionic or older, lxd comes preinstalled
as a debian package rather than a snap. Since the lxd versions are much
older than the snap, this behavior used to cause the lxd server offering
unsupported tls version during `juju bootstrap` therefore ending up with
a broken charm.

With this commit, we are using installing the lxd snap and running
lxd.migrate if lxd is installed as a debian package during the install
event handler to make sure we are using the snap everywhere.

Closes-Bug: 1988477

Signed-off-by: Mert Kırpıcı <mert.kirpici@canonical.com>